build: make CSS watching "incremental"
